requirements.txt
fastapi
uvicorn
tensorflow
pydantic
SQLAlchemy
transformers
numpy
google-auth
google-auth-oauthlib
google-api-python-client
scikit-learn
nlpaug
nltk
beautifulsoup4
langdetect